+++ mod-source-record-storage-server/src/main/java/org/folio/services/util/parser/lexeme/operand/BinaryOperand.java
@@ -4,6 +4,7 @@ import org.folio.services.util.parser.lexeme.Lexeme;
 import org.folio.services.util.parser.lexeme.Lexicon;
 
 import java.util.List;
+import java.util.Optional;
 
 /**
  * The interface for the binary operands. It has a key(field.suffix), operator, and right operand (value).
@@ -12,7 +13,7 @@ import java.util.List;
  */
 public interface BinaryOperand {
 
-  String getField();
+  Optional<String> getField();
 
   List<String> getBindingParams();
 }