+++ mod-source-record-storage-server/src/main/java/org/folio/services/util/parser/lexeme/operand/SubFieldBinaryOperand.java
@@ -4,6 +4,7 @@ import org.folio.services.util.parser.lexeme.Lexicon;
 
 import static java.lang.String.format;
 import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_EQUALS;
+import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_IS;
 import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_LEFT_ANCHORED_EQUALS;
 import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_NOT_EQUALS;
 
@@ -22,6 +23,8 @@ import static org.folio.services.util.parser.lexeme.Lexicon.BINARY_OPERATOR_NOT_
  *  010.a = '   55001156/M '    - simple equality
  *  010.a ^= '   55'            - left-anchored equality
  *  010.a not= '   55001156/M ' - not equals
+ *  010.a is 'present'          - check sub field for presence
+ *  010.a is 'absent'           - check sub field for absence
  */
 public class SubFieldBinaryOperand extends BinaryOperandLexeme {
 
@@ -36,7 +39,8 @@ public class SubFieldBinaryOperand extends BinaryOperandLexeme {
   @Override
   public String toSqlRepresentation() {
     String[] keyParts = getKey().split("\\.");
-    String iField = "\"" + "i" + keyParts[0] + "\"";
+    String field = keyParts[0];
+    String iField = "\"" + "i" + field + "\"";
     String subField = keyParts[1];
     StringBuilder stringBuilder = new StringBuilder()
       .append("(").append(iField).append(".\"subfield_no\" = '").append(subField).append("'")
@@ -48,6 +52,8 @@ public class SubFieldBinaryOperand extends BinaryOperandLexeme {
       return stringBuilder.append("= ?)").toString();
     } else if (BINARY_OPERATOR_NOT_EQUALS.equals(getOperator())) {
       return stringBuilder.append("<> ?)").toString();
+    } else if (BINARY_OPERATOR_IS.equals(getOperator())) {
+      return PresenceBinaryOperand.getSqlRepresentationForSubField(field, subField, value);
     }
     throw new IllegalArgumentException(format("Operator [%s] is not supported for the given SubField operand", getOperator().getSearchValue()));
   }